+++ src/main/java/com/caio/flowtrack_api/service/UserService.java
@@ -1,6 +1,8 @@
 package com.caio.flowtrack_api.service;
 
 import com.caio.flowtrack_api.entity.User;
+import com.caio.flowtrack_api.exception.DuplicateResourceException;
+import com.caio.flowtrack_api.exception.ResourceNotFoundException;
 import com.caio.flowtrack_api.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,6 +17,10 @@ public class UserService {
     private final UserRepository userRepository;
 
     public User create(User user) {
+        if(userRepository.existsByEmail(user.getEmail())){
+            throw new DuplicateResourceException("Email already exists");
+        }
+
         return userRepository.save(user);
     }
 
@@ -22,11 +28,15 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> findById(Long id) {
-        return userRepository.findById(id);
+    public User findById(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
     }
 
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User not found");
+        }
         userRepository.deleteById(id);
     }
 }
